.requirements.txt
 fastapi
  pydantic 
  pandas 
  llama-index 
  openai 
  uvicorn
python-dotenv
llama-index-experimental